+++ src/main/java/roomescape/dao/ReservationQueryingDAO.java
@@ -1,11 +1,11 @@
-package roomescape;
+package roomescape.dao;
 
 import java.util.List;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
-import roomescape.Time;
-import roomescape.Reservation;
+import roomescape.domain.Reservation;
+import roomescape.domain.Time;
 
 @Repository
 public class ReservationQueryingDAO {
