+++ src/main/java/roomescape/domain/Time.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
