+++ src/main/java/roomescape/service/TimeService.java
@@ -1,13 +1,15 @@
-package roomescape;
+package roomescape.service;
 
-import java.net.URI;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.ExceptionHandler;
+import roomescape.exception.NotFoundTimeException;
+import roomescape.dao.TimeQueryingDAO;
+import roomescape.dao.TimeUpdatingDAO;
+import roomescape.domain.Time;
 
 @Service
 public class TimeService {
